Split scripts into its own file again, now with Parser added
